backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "app"
authors = [
    {name = "froonn", email = "137333195+froonn@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.1.2)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
